+++ lib/components/decorated_navbar.dart
@@ -1,19 +1,15 @@
 part of "../persistent_bottom_nav_bar_v2.dart";
 
 class DecoratedNavBar extends StatelessWidget {
-  DecoratedNavBar({
+  const DecoratedNavBar({
     required this.child,
     super.key,
     this.decoration = const NavBarDecoration(),
-    ImageFilter? filter,
-    this.opacity = 1.0,
     this.height = kBottomNavigationBarHeight,
-  }) : filter = filter ?? ImageFilter.blur(sigmaX: 3, sigmaY: 3);
+  });
 
   final NavBarDecoration decoration;
-  final ImageFilter filter;
   final Widget child;
-  final double opacity;
   final double height;
 
   @override
@@ -21,11 +17,9 @@ class DecoratedNavBar extends StatelessWidget {
         borderRadius:
             decoration.borderRadius as BorderRadius? ?? BorderRadius.zero,
         child: BackdropFilter(
-          filter: filter,
+          filter: decoration.filter ?? ImageFilter.blur(sigmaX: 3, sigmaY: 3),
           child: DecoratedBox(
-            decoration: decoration.copyWith(
-              color: decoration.color?.withOpacity(opacity),
-            ),
+            decoration: decoration,
             child: SafeArea(
               top: false,
               right: false,