+++ lib/styles/style_1_bottom_navbar.dart
@@ -47,8 +47,6 @@ class Style1BottomNavBar extends StatelessWidget {
   @override
   Widget build(BuildContext context) => DecoratedNavBar(
         decoration: navBarDecoration,
-        filter: navBarConfig.selectedItem.filter,
-        opacity: navBarConfig.selectedItem.opacity,
         height: navBarConfig.navBarHeight,
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceAround,