+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Use `NavBarOverlap.none()` as the default for `navBarOverlap`
+- Removed ItemConfig.opacity. Use the opacity of NavBarDecoration.color instead
+- Removed ItemConfig.filter. Use NavBarDecoration.filter instead
+- Removed default value of NavBarDecoration.filter to allow no filter at all
 
 ## [5.0.0] - 2024-03-13
 ### Added